# 01-Documents/API-SPEC-001-openapi.yaml
openapi: 3.0.3
info:
  title: RIS Invoice Discounting Platform API
  version: 1.0.0
  description: |
    REST API for the RIS Invoice Discounting and Early Payment Platform, Uganda.
    All endpoints require Bearer JWT authentication unless marked public.
    Role-based access is enforced at route and service layer independently.
  contact:
    name: RIS Engineering
  license:
    name: Proprietary
servers:
  - url: https://api.mms.co.ug/v1
    description: Production
  - url: https://staging-api.mms.co.ug/v1
    description: Staging
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
          example: CREDIT_LIMIT_EXCEEDED
        message:
          type: string
          example: Invoice face value exceeds buyer credit limit
        details:
          type: object
          description: Field-level error details for validation failures
...
    UUID:
      type: string
      format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
...
    PaymentProvider:
      type: string
      enum: [MTN_MOMO, AIRTEL, EFT]

    # ─── Auth Schemas ───────────────────────────────────────────────────────
...
    SupplierResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        company_name:
          type: string
        registration_number:
          type: string
        kyc_status:
          type: string
          enum: [pending, documents_submitted, under_review, approved, rejected]
        preferred_payment_method:
          $ref: '#/components/schemas/PaymentProvider'
        created_at:
          type: string
          format: date-time
        sanctions_flag:
          type: boolean

    # ─── Invoice Schemas ────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Not authenticated or token expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error_code: UNAUTHORIZED
            message: Authentication required
    Forbidden:
      description: Authenticated but insufficient role or accessing another user's resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error_code: FORBIDDEN
            message: You do not have permission to access this resource
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error — no stack trace exposed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error_code: INTERNAL_ERROR
            message: An unexpected error occurred. Please contact support.

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /suppliers/{id}:
    get:
      summary: Get supplier by ID
      tags: [Suppliers]
      x-required-role: supplier (own only), credit_officer, management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Supplier details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /admin/suppliers:
    get:
      summary: List all suppliers with filters
      tags: [Admin]
      x-required-role: credit_officer, compliance_officer, management
      parameters:
        - name: kyc_status
          in: query
          schema:
            type: string
            enum: [pending, documents_submitted, under_review, approved, rejected]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Paginated list of suppliers
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SupplierResponse'
                  total:
                    type: integer
                  page:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─── Invoices ──────────────────────────────────────────────────────────────
...